Updated integration test instance subnet.
The subnet now used lives in a vcn inside the kubernetes-test
compartment. Previously, the subnet was in bristol-cloud, thus the
test user needed bristol-cloud permissions.

Also updated the scripts such that the test bnary/config is not in the
same location as the volumes we are trying to mount/unmount, as this
was causing sporadic failures in the CI.